server/requirements.txt
openenv-core[core]>=0.2.1
uvicorn>=0.30.0
